Add distribution sampling routines
As per private discussion, I used a log-normal distribution instead
of KDE (which was originally used in the paper).

A summary of my comments in favour of this solution:
- KDE: computational complexity is 1 exponentiation. However, it
  requires O(n) memory.
- GMM: would have been a good alternative, using O(k) memory k << m.
- Log-normal: requires O(1) memory, has similar computational complexity
  of the previous, and to my intuition it will approximate well our
  data. More experiments are required to test this.

Bogus distribution parameters are in use for now.

NOTE: in the future we may want to specify distribution parameters in
some conf file (even if at compilation time)