Pin nightly version against clippy
Fixes #16. This way we don't have to recompile clippy against the latest Rust
nightly every time it updates. Every few weeks/ months we should update the
clippy and Rust version. This is less than ideal, but it was the best solution I
found (via https://dev.to/cad97/great-rust-ci-1fk6 ).